Move ReflexionHelper to right package
