+++ tests/Helper/ReflexionHelper.php
@@ -25,11 +25,23 @@
 *  International Registered Trademark & Property of PrestaShop SA
 */
 
-namespace PrestaShop\PrestaShop\Tests\TestCase;
+namespace PrestaShop\PrestaShop\Tests\Helper;
 
 use ReflectionClass;
 use PHPUnit_Framework_TestCase;
 
+/**
+ * Class ReflexionHelper
+ * @package PrestaShop\PrestaShop\Tests\TestCase
+ *
+ * Provides utilities to access private or protected properties inside classes.
+ *
+ * Please be careful with these feature, testing private fields or methods is often a bad smell :
+ *      - you may be testing your implementation details rather than the behaviour of your code.
+ *      - if this is because the class under test is too long and complicated, you should defintely consider breaking this class into several smaller ones.
+ *
+ * In the end, this kind of features is here just for convenience to be able to test quickly dirty legacy code.
+ */
 class ReflexionHelper extends PHPUnit_Framework_TestCase
 {
